requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.3
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.1.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
